[erts] Priority Messages - Increase amount of monitor/link flags
The 16-bit monitor/link flags field is almost full and we need more flags.
The 16-bit type field use very few bits so we merge the 16-bit type field
and the 16-bit flags field into one new 32-bit flags field. By this we make
a lot of new flag bits available without increasing the size of monitor and
link structures.

+++ erts/emulator/beam/bif.h
@@ -585,7 +585,7 @@ erts_reschedule_bif(Process *proc,
 
 #endif /* ERTS_GLB_INLINE_INCL_FUNC_DEF */
 
-Uint16 erts_monitor_opts(Eterm opts, Eterm *tag);
+Uint32 erts_monitor_opts(Eterm opts, Eterm *tag);
 
 #include "erl_bif_table.h"
